tests: Move filesystem tests into separate directory
The original fs_test.py was more 3k lines and very hard to
navigate. This splits the tests in the same way the fs plugin
itself is split.
