+++ tests/fs_tests/__init__.py
@@ -1,0 +1,15 @@
+import sys
+sys.path.append("..")
+
+import overrides_hack
+
+from .exfat_test import *
+from .ext_test import *
+from .f2fs_test import *
+from .generic_test import *
+from .mount_test import *
+from .nilfs_test import *
+from .ntfs_test import *
+from .reiserfs_test import *
+from .vfat_test import *
+from .xfs_test import *
